Membuat Variabel di Finance menjadi variabel global dan mengatur logo profile
