+++ lib/GlobalVariabel.dart
@@ -1,5 +1,4 @@
-
-
+import 'package:flutter/foundation.dart'; 
 class GlobalData {
   static double money = 1000000.0;
 }
@@ -30,5 +29,40 @@ class Balance {
   void deductMoney(int balance) {
     this.balance -= balance;
   }
+}
+
+
+class Coin {
+  final ValueNotifier<double> jumlah;
+  final double harga;
+  final String symbol;
+
+  Coin({required this.jumlah, required this.harga, required this.symbol});
+}
+
+class GlobalVariables {
+  static final Coin bitcoin = Coin(
+    symbol: 'BTC',
+    harga: 29850.0,
+    jumlah: ValueNotifier<double>(0),
+  );
+
+  static final Coin ethereum = Coin(
+    symbol: 'ETH',
+    harga: 10560.0,
+    jumlah: ValueNotifier<double>(0),
+  );
+
+  static final Coin litecoin = Coin(
+    symbol: 'LTC',
+    harga: 3676.0,
+    jumlah: ValueNotifier<double>(0),
+  );
+
+  static final Coin ripple = Coin(
+    symbol: 'XRP',
+    harga: 5240.0,
+    jumlah: ValueNotifier<double>(0),
+  );
 }
