+++ pubspec.yaml
@@ -93,7 +93,7 @@ flutter:
     - assets/reward.png
     - assets/setting.png
     - assets/tanggal.png
-
+    - assets/Profil.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
